107e108e109e110_refatorados

+++ Python3/exe107.py
@@ -3,11 +3,11 @@
 import sys,os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'modulos'))) # Adiciona a pasta 'modulos' ao sys.path
 
-from moedas import aumentar,diminuir,dobro,metade,moeda
+from moedas import aumentar,diminuir,dobro,metade
 from helpers import título
 
 título('Treinamento de Módulos',2)
-num = int(input('Digite o preço: R$ '))
+num = float(input('Digite o preço: R$ '))
 
 print(f'A metade de R$ {num} é R${metade(num)}.')
 print(f'O dobro de R$ {num} é R$ {dobro(num)}.')
